fix: SEO, OG image previews, port and sitemap fixes
- Convert homepage to SSR Server Component for Googlebot crawlability
- Extract client code to components/home/home-client.tsx with initialData props
- Fix sitemap: was returning 0 article URLs (Array.isArray bug on paginated response)
- Add generateStaticParams + revalidate to article pages
- Add BreadcrumbList JSON-LD and WebSite SearchAction JSON-LD
- Add metadata layouts for contributors and github pages
- Fix OG image preview: sanitize localhost URLs, use relative /api/og path
- Fix backend upload-controller: use API_URL env var instead of localhost:5000
- Add OG image dimensions (1200x630) to layout and article metadata
- Change port 6000 -> 8080 (6000 is browser-blocked unsafe port)
- Fix next.config.mjs image remote pattern port 5000 -> 8080
- Reformat prisma/schema.prisma (prisma db pull normalization)
