+++ backend/prisma/schema.prisma
@@ -7,67 +7,86 @@ datasource db {
   url      = env("DATABASE_URL")
 }
 
-model User {
-  id            String    @id @default(cuid())
-  walletAddress String    @unique
-  email         String?   @unique
-  name          String?
-  role          Role      @default(WRITER)
-  avatarUrl     String?
-  bio           String?
-  articles      Article[]
-  comments      Comment[]
-  likes         Like[]
-  bookmarks     Bookmark[]
-  createdAt     DateTime  @default(now())
-  updatedAt     DateTime  @updatedAt
-
-  @@index([walletAddress])
-}
-
 model Article {
-  id            String       @id @default(cuid())
+  id            String         @id
   title         String
-  slug          String       @unique
+  slug          String         @unique
   description   String
   content       String
   category      String
-  featuredImage String?
-  tags          TagOnArticle[]
-  author        User         @relation(fields: [authorId], references: [id])
   authorId      String
-  likes         Like[]
-  comments      Comment[]
-  bookmarks     Bookmark[]
-  views         ArticleView[]
-  viewCount     Int          @default(0)
-  uniqueViews   Int          @default(0)
-  publishedAt   DateTime     @default(now())
-  status        ArticleStatus @default(DRAFT)
-  featured      Boolean      @default(false)
-  createdAt     DateTime     @default(now())
-  updatedAt     DateTime     @updatedAt
+  publishedAt   DateTime       @default(now())
+  status        ArticleStatus  @default(DRAFT)
+  featured      Boolean        @default(false)
+  createdAt     DateTime       @default(now())
+  updatedAt     DateTime
+  uniqueViews   Int            @default(0)
+  viewCount     Int            @default(0)
+  featuredImage String?
+  User          User           @relation(fields: [authorId], references: [id])
+  ArticleView   ArticleView[]
+  Bookmark      Bookmark[]
+  Comment       Comment[]
+  Like          Like[]
+  TagOnArticle  TagOnArticle[]
 
-  @@index([viewCount])
-  @@index([uniqueViews])
   @@index([authorId])
-  @@index([status])
-  @@index([publishedAt])
-  @@index([featured])
-  @@index([createdAt])
-  @@index([status, publishedAt])
   @@index([authorId, status])
+  @@index([createdAt])
+  @@index([featured])
+  @@index([publishedAt])
+  @@index([status])
+  @@index([status, publishedAt])
+  @@index([uniqueViews])
+  @@index([viewCount])
+}
+
+model ArticleView {
+  id          String   @id
+  articleId   String
+  userId      String?
+  visitorId   String
+  ipAddress   String?
+  userAgent   String?
+  referrer    String?
+  country     String?
+  city        String?
+  device      String?
+  browser     String?
+  os          String?
+  viewedAt    DateTime @default(now())
+  timeOnPage  Int?
+  scrollDepth Int?
+  Article     Article  @relation(fields: [articleId], references: [id], onDelete: Cascade)
+
+  @@unique([articleId, visitorId, viewedAt])
+  @@index([articleId])
+  @@index([viewedAt])
+  @@index([visitorId])
+}
+
+model Bookmark {
+  id        String   @id
+  userId    String
+  articleId String
+  createdAt DateTime @default(now())
+  Article   Article  @relation(fields: [articleId], references: [id])
+  User      User     @relation(fields: [userId], references: [id])
+
+  @@unique([userId, articleId])
+  @@index([articleId])
+  @@index([userId, createdAt])
 }
 
 model Comment {
-  id        String   @id @default(cuid())
+  id        String        @id
   body      String
-  article   Article  @relation(fields: [articleId], references: [id])
   articleId String
-  author    User     @relation(fields: [authorId], references: [id])
   authorId  String
   status    CommentStatus @default(PENDING)
-  createdAt DateTime @default(now())
+  createdAt DateTime      @default(now())
+  Article   Article       @relation(fields: [articleId], references: [id])
+  User      User          @relation(fields: [authorId], references: [id])
 
   @@index([articleId])
   @@index([authorId])
@@ -76,83 +95,58 @@ model Comment {
 }
 
 model Like {
-  id        String   @id @default(cuid())
-  article   Article  @relation(fields: [articleId], references: [id])
+  id        String   @id
   articleId String
-  user      User     @relation(fields: [userId], references: [id])
   userId    String
   createdAt DateTime @default(now())
+  Article   Article  @relation(fields: [articleId], references: [id])
+  User      User     @relation(fields: [userId], references: [id])
 
   @@unique([articleId, userId])
-  @@index([userId])
-  @@index([createdAt])
   @@index([articleId, createdAt])
-}
-
-model Bookmark {
-  id        String   @id @default(cuid())
-  user      User     @relation(fields: [userId], references: [id])
-  userId    String
-  article   Article  @relation(fields: [articleId], references: [id])
-  articleId String
-  createdAt DateTime @default(now())
-
-  @@unique([userId, articleId])
-  @@index([userId, createdAt])
-  @@index([articleId])
+  @@index([createdAt])
+  @@index([userId])
 }
 
 model Subscription {
-  id        String   @id @default(cuid())
+  id        String   @id
   email     String   @unique
   topics    String[]
   createdAt DateTime @default(now())
 }
 
 model Tag {
-  id     String          @id @default(cuid())
-  name   String          @unique
-  slug   String          @unique
-  posts  TagOnArticle[]
+  id           String         @id
+  name         String         @unique
+  slug         String         @unique
+  TagOnArticle TagOnArticle[]
 }
 
 model TagOnArticle {
-  article   Article @relation(fields: [articleId], references: [id])
   articleId String
-  tag       Tag     @relation(fields: [tagId], references: [id])
   tagId     String
+  Article   Article @relation(fields: [articleId], references: [id])
+  Tag       Tag     @relation(fields: [tagId], references: [id])
 
   @@id([articleId, tagId])
 }
 
-model ArticleView {
-  id              String   @id @default(cuid())
-  article         Article  @relation(fields: [articleId], references: [id], onDelete: Cascade)
-  articleId       String
-  userId          String?  // Optional - for logged-in users
-  visitorId       String   // Browser fingerprint or session ID
-  ipAddress       String?
-  userAgent       String?
-  referrer        String?
-  country         String?
-  city            String?
-  device          String?  // mobile, desktop, tablet
-  browser         String?
-  os              String?
-  viewedAt        DateTime @default(now())
-  timeOnPage      Int?     // seconds spent on page
-  scrollDepth     Int?     // percentage scrolled
+model User {
+  id            String     @id
+  walletAddress String     @unique
+  name          String?
+  role          Role       @default(WRITER)
+  avatarUrl     String?
+  bio           String?
+  createdAt     DateTime   @default(now())
+  updatedAt     DateTime
+  email         String?    @unique
+  Article       Article[]
+  Bookmark      Bookmark[]
+  Comment       Comment[]
+  Like          Like[]
 
-  @@unique([articleId, visitorId, viewedAt])
-  @@index([articleId])
-  @@index([visitorId])
-  @@index([viewedAt])
-}
-
-enum Role {
-  ADMIN
-  EDITOR
-  WRITER
+  @@index([walletAddress])
 }
 
 enum ArticleStatus {
@@ -165,4 +159,10 @@ enum CommentStatus {
   PENDING
   APPROVED
   REJECTED
+}
+
+enum Role {
+  ADMIN
+  EDITOR
+  WRITER
 }
