+++ frontend/next.config.mjs
@@ -28,7 +28,7 @@ const config = {
       {
         protocol: 'http',
         hostname: 'localhost',
-        port: '5000',
+        port: '8080',
         pathname: '/uploads/**',
       },
       {
